更新configmanager实现 进行高频调用点优化 更新MCP23S17和GTX312L模块 UART HAL添加硬件流控支持 注册主板引脚

+++ src/protocol/mai2serial/mai2serial.cpp
@@ -268,8 +268,8 @@ void Mai2Serial::show_status_info() {
     }
     
     char info[128];
-    snprintf(info, sizeof(info), "Status: %s, Baud: %u", 
-             status_str, config_.baud_rate);
+    snprintf(info, sizeof(info), "Status: %s, Baud: %lu", 
+             status_str, static_cast<unsigned long>(config_.baud_rate));
     send_response(info);
 }
